getting all clients by id api route sending info back to client -- no validations yet
